% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleProps-package.R
\docType{package}
\name{ResampleProps-package}
\alias{ResampleProps}
\alias{ResampleProps-package}
\title{ResampleProps: Resample One and Two Sample Means and Proportions}
\description{
THe package contains four commands designed to create the sampling distributions of means and proportions in one sample and in two independent samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertwwalker/ResampleProps}
  \item Report bugs at \url{https://github.com/robertwwalker/ResampleProps/issues}
}

}
\author{
\strong{Maintainer}: Robert Walker \email{rwalker@willamette.edu}

}
\keyword{internal}
